+++ futures/src/client.rs
@@ -1,7 +1,9 @@
 use futures::{Future, Poll};
 use lapin::{confirmation::Confirmation, Connect as LapinConnect, Connection};
 
-use crate::{uri::AMQPUri, Channel, ConfirmationFuture, ConnectionProperties, Error};
+use crate::{
+    tcp::Identity, uri::AMQPUri, Channel, ConfirmationFuture, ConnectionProperties, Error,
+};
 
 /// Connect to a server and create channels
 #[derive(Clone)]
@@ -12,12 +14,30 @@ pub struct Client {
 impl Client {
     /// Connect to an AMQP Server
     pub fn connect(uri: &str, options: ConnectionProperties) -> ClientFuture {
-        Connect::connect(uri, options)
+        Connect::connect(uri, options, None)
+    }
+
+    /// Connect to an AMQP Server
+    pub fn connect_with_identity(
+        uri: &str,
+        options: ConnectionProperties,
+        identity: Identity<'_, '_>,
+    ) -> ClientFuture {
+        Connect::connect(uri, options, Some(identity))
     }
 
     /// Connect to an AMQP Server
     pub fn connect_uri(uri: AMQPUri, options: ConnectionProperties) -> ClientFuture {
-        Connect::connect(uri, options)
+        Connect::connect(uri, options, None)
+    }
+
+    /// Connect to an AMQP Server
+    pub fn connect_uri_with_identity(
+        uri: AMQPUri,
+        options: ConnectionProperties,
+        identity: Identity<'_, '_>,
+    ) -> ClientFuture {
+        Connect::connect(uri, options, Some(identity))
     }
 
     /// Return a future that resolves to a `Channel` once the method succeeds
@@ -66,17 +86,29 @@ impl From<Confirmation<Connection>> for ClientFuture {
 /// Trait providing a method to connect to an AMQP server
 pub trait Connect {
     /// Connect to an AMQP server
-    fn connect(self, options: ConnectionProperties) -> ClientFuture;
+    fn connect(
+        self,
+        options: ConnectionProperties,
+        identity: Option<Identity<'_, '_>>,
+    ) -> ClientFuture;
 }
 
 impl Connect for AMQPUri {
-    fn connect(self, options: ConnectionProperties) -> ClientFuture {
-        LapinConnect::connect(self, options).into()
+    fn connect(
+        self,
+        options: ConnectionProperties,
+        identity: Option<Identity<'_, '_>>,
+    ) -> ClientFuture {
+        LapinConnect::connect(self, options, identity).into()
     }
 }
 
 impl Connect for &str {
-    fn connect(self, options: ConnectionProperties) -> ClientFuture {
-        LapinConnect::connect(self, options).into()
+    fn connect(
+        self,
+        options: ConnectionProperties,
+        identity: Option<Identity<'_, '_>>,
+    ) -> ClientFuture {
+        LapinConnect::connect(self, options, identity).into()
     }
 }